+++ util/aescbc/common_test.go
@@ -1,4 +1,4 @@
-package util
+package aescbc
 
 import (
 	"log"
